Solution for Requirement 2 and 3: - Why you chose the strategy you did A) In models.py for NearEarthObject, added is_potentially_hazardous_asteroid field to NEODatabase,as that is one of the filter criterias for requirements 2/3
In search.py, populated the Filter class also added code to return NEO or ORBITS based on the suitable return_types

- The data structures you used to implement that search case and why
A)
In search.py - filter class, am looking into the field property. Based on the requirements when field = NearEarthObject, the objects to compare are diameter and is_hazardous, similarly when the field = OrbitPath, the only thing to compare is distance parameter.
In Neosearcer.py looked added the logic to return the types the user asked for based on the return_type property